requirements.txt
PyPDF2
python-docx
pypiwin32
